Problem 8 : still on b but Im getting close.
Added display.go, started working with the pterm package. Created methods for visualization of the exercise.
In nodes.go, renamed node and NodeState to walker and team respectively and converted them to structs.
In partb.go, applied appropriate changes to accommodate the aforementioned change in functionality.
Updated go.mod for the addition of pterm and go-echarts.